refactor: start incremental cdc during registration
